more linting / style cleanup
