+++ src/hooks/useFindMatches.ts
@@ -8,14 +8,14 @@ import { SerializedTabState, TabState } from '../types/tab.types';
 import { serializeMatchesObj } from '../utils/htmlUtils';
 import { searchAndHighlight } from '../utils/matchUtils/highlightUtils';
 import { sendMsgToBackground } from '../utils/messageUtils/sendMessageToBackground';
-import { scrollToElement } from '../utils/scrollUtils';
+import scrollToElement from '../utils/scrollUtils';
 
 type UpdateHighlightsOptions = {
   prevIndex?: number;
   endOfTab?: boolean;
 };
 
-export const useFindMatches = () => {
+export default function useFindMatches() {
   const { totalMatchesCount, globalMatchIdx } = useContext(LayoverContext);
   const { tabStateContext, setTabStateContext } = useContext(TabStateContext);
 
@@ -219,4 +219,4 @@ export const useFindMatches = () => {
     previousMatch,
     updateHighlights,
   };
-};
+}
