+++ src/utils/backgroundUtils.ts
@@ -1,9 +1,10 @@
 // src/utils/backgroundUtils.ts
 
 // import { store } from '../background/background';
-import { WindowStore, updateStore } from '../background/store';
 
+import { updateStore } from '../background/store';
 import { LayoverPosition } from '../types/Layover.types';
+import { WindowStore } from '../types/Store.types';
 import {
   HighlightMsg,
   RemoveAllHighlightMatchesMsg,
