+++ src/utils/scrollUtils.ts
@@ -1,4 +1,5 @@
 // src/utils/scrollUtils.ts
-export function scrollToElement(element: HTMLElement) {
+
+export default function scrollToElement(element: HTMLElement) {
   element.scrollIntoView({ behavior: 'smooth', block: 'center' });
 }
